+++ docs/CHANGELOG.rst
@@ -32,6 +32,8 @@ Changed
 - **BACKWARD INCOMPATIBLE:** Remove filtering of genes with low expression
   in PCA analysis
 - Expand error messages in ``rose2`` process
+- Check for errors during download of FASTQ files and use
+  ``resolwebio/utils:1.3.0`` Docker image in import SRA process
 
 Fixed
 -----